log writer refactoring
