+++ src/communication/network_interface_sr.h
@@ -33,7 +33,7 @@
 
 #include "file_io.h"
 #include "log_comm.h"
-#include "log_common_impl.h"
+#include "log_writer.h"
 #include "method_scan.h"
 #include "thread_compat.hpp"
