+++ src/executables/util_common.c
@@ -36,6 +36,7 @@
 #include "porting.h"
 #include "message_catalog.h"
 #include "log_common_impl.h"
+#include "log_writer.h"
 #include "mprec.h"
 #include "system_parameter.h"
 #include "environment_variable.h"
